% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_theme.R
\name{scale_fill_veraz}
\alias{scale_fill_veraz}
\title{Veraz scale colors for ggplot2}
\usage{
scale_fill_veraz(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{the palette to be used, options are \code{c('main', 'secondary', 'grey')}}

\item{discrete}{boolean, the scale is discrete? \code{TRUE} if it is.}

\item{reverse}{boolean, represents the order of the palette}

\item{...}{Named argument to modify the \code{scale_color_} fucntion}
}
\description{
Veraz scale colors for ggplot2
}
