% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_theme.R
\name{veraz_pal}
\alias{veraz_pal}
\title{Varaz color palettes}
\usage{
veraz_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{the palette to be used, options are \code{c('main', 'secondary', 'grey')}}

\item{reverse}{boolean, represents the order of the palette}

\item{...}{Named argument to modify the palette}
}
\description{
Varaz color palettes
}
